refactoring: search_task.dart 파일 생성
